+++ translation/chinese/08-數字和字串/01-數字/練習/01-計算/解答/main.go
