+++ translation/chinese/08-數字和字串/01-數字/練習/02-浮點數/解答/main.go
@@ -11,7 +11,7 @@ package main
 import "fmt"
 
 func main() {
-	// Below solutions are correct:
+	// 以下正確:
 	x := 5. / 2
 	// x := 5 / 2.
 	// x := float64(5) / 2
