+++ translation/chinese/08-數字和字串/01-數字/練習/04-遞增遞減/main.go
@@ -1,0 +1,34 @@
+// Copyright © 2018 Inanc Gumus
+// Learn Go Programming Course
+// License: https://creativecommons.org/licenses/by-nc-sa/4.0/
+//
+// For more tutorials  : https://learngoprogramming.com
+// In-person training  : https://www.linkedin.com/in/inancgumus/
+// Follow me on twitter: https://twitter.com/inancgumus
+
+package main
+
+// ---------------------------------------------------------
+// 練習：遞增遞減
+//
+//  1. 將 `counter` 增加 5 次
+//  2. 將 `factor` 減少 2 次
+//  3. 輸出 counter 和 factor 的乘積
+//
+// 限制條件：
+//  只能使用遞增與遞減語句
+//
+// 預期輸出
+//  -75
+// ---------------------------------------------------------
+
+func main() {
+	// 不用改這裡
+	counter, factor := 45, 0.5
+
+	// 在下方寫程式碼
+	// ...
+
+	// 最後，移除以下程式碼
+	_, _ = counter, factor
+}
