+++ translation/chinese/08-數字和字串/01-數字/練習/04-遞增遞減/解答/main.go
