+++ translation/chinese/08-數字和字串/01-數字/練習/05-操作計數器/main.go
@@ -1,0 +1,38 @@
+// Copyright © 2018 Inanc Gumus
+// Learn Go Programming Course
+// License: https://creativecommons.org/licenses/by-nc-sa/4.0/
+//
+// For more tutorials  : https://learngoprogramming.com
+// In-person training  : https://www.linkedin.com/in/inancgumus/
+// Follow me on twitter: https://twitter.com/inancgumus
+
+package main
+
+import "fmt"
+
+// ---------------------------------------------------------
+//練習:操作計數器
+//
+//  1. 寫出最簡單的一行程式碼，讓 counter 變數增加 1。
+//
+//  2. 寫出最簡單的一行程式碼，讓 counter 變數減少 1。
+//
+//  3. 寫出最簡單的一行程式碼，讓 counter 變數增加 5。
+//
+//  4. 寫出最簡單的一行程式碼，讓 counter 變數乘以 10。
+//
+//  5. 寫出最簡單的一行程式碼，讓 counter 變數除以 5。
+//
+// 預期輸出
+//  10
+// ---------------------------------------------------------
+
+func main() {
+	// 不用改這裡
+	var counter int
+
+	// 在此輸入程式碼
+
+	// 不用改這裡
+	fmt.Println(counter)
+}
