+++ translation/chinese/08-數字和字串/01-數字/練習/05-操作計數器/解答/main.go
