+++ translation/chinese/08-數字和字串/01-數字/練習/06-簡化敘述/解答/main.go
