+++ translation/chinese/08-數字和字串/01-數字/練習/07-圓面積/解答/main.go
@@ -24,7 +24,7 @@ func main() {
 	fmt.Printf("radius: %g -> area: %.2f\n",
 		radius, area)
 
-	// ALTERNATIVE:
-	// math.Pow calculates the power of a float number
+	// 替代方法：
+	// math.Pow 用來計算浮點數的次方
 	// area = math.Pi * math.Pow(radius, 2)
 }
